requirements.txt
deepgram-sdk==3.2.5
python-dotenv==1.0.1
pyaudio==0.2.14
numpy==1.26.4
scipy==1.13.0
pick==2.2.0
supabase-py-async==2.5.6
openai==1.17.1